+++ retz-client/src/main/java/io/github/retz/cli/CommandSchedule.java
@@ -52,6 +52,8 @@ public class CommandSchedule implements SubCommand {
     private String appName;
     @Parameter(names = {"--prio", "--priority"}, description = "Job priority")
     int priority = 0;
+    @Parameter(names = {"--attributes"}, description = "Attributes of the job (free format, depends on the planner)")
+    String attributes = null;
     @Parameter(names = {"-N", "--name"}, description = "Human readable job name")
     String name;
     @Parameter(names="--tags", description = "Tags separated by commas (e.g. 'a,b,c')")
@@ -75,6 +77,7 @@ public class CommandSchedule implements SubCommand {
         job.setPriority(priority);
         job.setName(name);
         job.addTags(tags);
+        job.setAttributes(attributes);
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
                 .setAuthenticator(fileConfig.getAuthenticator())